Implement `dora-rover` logic
This commit introduce a lot of the logic required to run a real environment
rover.

It changes the basic localisation from euler angle to quaternion as it is
more robust.

It also transform the planner from `hybrid_astar` to `frenet`.

It add a bit of visualisation information concerning the position and control output.

It removes the precalculated speed to an estimated speed.